add tests

+++ src/byte_fns.rs
@@ -1,7 +1,7 @@
 use crate::Q;
 
 // Q=3329
-pub(crate) fn compress(d:usize, u:&[u16;256]) -> [u16;256] {
+pub(crate) fn compress(d:u8, u:&[u16;256]) -> [u16;256] {
     let q = Q as u32;
     let m: u32 = 1<<d;
     // round(i/j) = floor((i+floor(j/2))/j)
@@ -10,7 +10,7 @@ pub(crate) fn compress(d:usize, u:&[u16;256]) -> [u16;256] {
 }
 
 // Q=3329
-pub(crate) fn decompress(d:usize, u:&[u16;256]) -> [u16;256] {
+pub(crate) fn decompress(d:u8, u:&[u16;256]) -> [u16;256] {
     let m: u32 = 1<<d;
     // round(i/j) = floor((i+floor(j/2))/j)
     // compute round((q*y)/2^d) -> round((Q*u[i])/m)
@@ -39,7 +39,7 @@ pub(crate) fn byte_encode<const LEN:usize>(d:u8, f:&[u16;256]) -> [u8;LEN] {
     b
 }
 
-pub(crate) fn byte_decode(d:usize, b:&[u8]) -> [u16;256] {
+pub(crate) fn byte_decode(d:u8, b:&[u8]) -> [u16;256] {
     // compute m
     let m: u16 = {if d==12 {Q} else {1<<d}};
